[ADD] 출석 체크 API 적용 완료

+++ domain/src/main/java/kr/co/domain/model/UserId.kt
@@ -1,5 +1,5 @@
 package kr.co.domain.model
 
 data class UserId(
-    val useId: Int
+    val useId: Long
 )
